Converted a bunch of signed int32_t into unsigned because it was causing a lot of needless sign extension on x64 and possibly other platforms. Merged the decompression bitset indices into a single track index. Fixes #143.

+++ includes/acl/decompression/default_output_writer.h
@@ -45,25 +45,25 @@ namespace acl
 			ACL_ASSERT(num_transforms != 0, "Transforms array cannot be empty");
 		}
 
-		void write_bone_rotation(int32_t bone_index, const Quat_32& rotation)
+		void write_bone_rotation(uint16_t bone_index, const Quat_32& rotation)
 		{
 			ACL_ASSERT(bone_index < m_num_transforms, "Invalid bone index. %u >= %u", bone_index, m_num_transforms);
 			m_transforms[bone_index].rotation = rotation;
 		}
 
-		void write_bone_translation(int32_t bone_index, const Vector4_32& translation)
+		void write_bone_translation(uint16_t bone_index, const Vector4_32& translation)
 		{
 			ACL_ASSERT(bone_index < m_num_transforms, "Invalid bone index. %u >= %u", bone_index, m_num_transforms);
 			m_transforms[bone_index].translation = translation;
 		}
 
-		void write_bone_scale(int32_t bone_index, const Vector4_32& scale)
+		void write_bone_scale(uint16_t bone_index, const Vector4_32& scale)
 		{
 			ACL_ASSERT(bone_index < m_num_transforms, "Invalid bone index. %u >= %u", bone_index, m_num_transforms);
 			m_transforms[bone_index].scale = scale;
 		}
 
 		Transform_32* m_transforms;
-		int32_t m_num_transforms;
+		uint16_t m_num_transforms;
 	};
 }
